refactor(pages): structure fix, moving default section from templates to layouts
