+++ src/layouts/index.ts
@@ -1,0 +1,4 @@
+import { DefaultSection } from './DefaultSection';
+import { Meta } from './Meta';
+
+export { DefaultSection, Meta };
