+++ src/pages/about.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image';
 
 import { NextLinkBox } from '@/components';
-import { Meta } from '@/layouts/Meta';
-import { DefaultSection, Main } from '@/templates';
+import { DefaultSection, Meta } from '@/layouts';
+import { Main } from '@/templates';
 
 const linkBoxData = [
   {
