+++ src/templates/index.ts
@@ -1,4 +1,3 @@
-import { DefaultSection } from './DefaultSection';
 import { Main } from './Main';
 
-export { DefaultSection, Main };
+export { Main };
